requirements.txt
astroid==2.15.0
boto3==1.26.104
botocore==1.29.104
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Deprecated==1.2.13
dill==0.3.6
Flask==2.2.3
flatten-json==0.1.13
googlemaps==4.10.0
greenlet==2.0.2
gunicorn==20.0.4
idna==3.4
importlib-metadata==6.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
networkx==2.8.8
numpy==1.24.2
pandas==1.5.3
platformdirs==3.1.1
pylint==2.17.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-tsp==0.3.1
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==2.0.6
tabulate==0.8.10
tomli==2.0.1
tomlkit==0.11.6
tsplib95==0.7.1
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.14.0
Flask-Session==0.4.0